chagned component structure
