+++ src/components/Movie/styles.css
@@ -1,0 +1,3 @@
+.img-placeholder {
+    
+}
